All exercises completed and tests passed

+++ controllers/blogs.js
@@ -2,12 +2,19 @@ const blogsRouter = require("express").Router();
 const Blog = require("../models/blog");
 
 blogsRouter.get("/", async (request, response) => {
-  const blogs = await Blog.find({});
+  const blogs = await Blog.find({}).populate("user", {
+    username: 1,
+    name: 1,
+  });
 
   response.json(blogs);
 });
 
 blogsRouter.post("/", async (request, response) => {
+  if (!request.token || !request.user) {
+    return response.status(401).send({ message: "Unauthorized" });
+  }
+
   const body = request.body;
 
   const blog = new Blog({
@@ -15,6 +22,7 @@ blogsRouter.post("/", async (request, response) => {
     author: body.author,
     url: body.url,
     likes: body.likes || 0,
+    user: request.user._id,
   });
 
   const newBlog = await blog.save();
@@ -23,6 +31,17 @@ blogsRouter.post("/", async (request, response) => {
 });
 
 blogsRouter.delete("/:id", async (request, response) => {
+  const { token, user } = request;
+  if (!token || !user) {
+    return response.status(401).send({ message: "Unauthorized" });
+  }
+
+  const blog = await Blog.findById(request.params.id);
+
+  if (blog.user.toString() !== user._id.toString()) {
+    return response.status(401).send({ message: "Unauthorized" });
+  }
+
   await Blog.findByIdAndDelete(request.params.id);
 
   response.send(204).end();
@@ -43,7 +62,7 @@ blogsRouter.put("/:id", async (request, response) => {
   });
 
   if (!updatedBlog) response.send(400).end();
-  
+
   response.json(updatedBlog);
 });
